# Algorithms Different sorts of algorithms (search and sorting)
I am using BubbleSort. Bubble Sort is the simplest sorting algorithm that works by repeatedly swapping the adjacent elements if they are in wrong order.

I am also using LinearSearch. Linear search is used on a collections of items. It relies on the technique of traversing a list from start to end by exploring properties of all the elements that are found on the way.